+++ packages/core/src/components/CenterAlignedLoader/CenterAlignedLoader.stories.tsx
@@ -1,0 +1,5 @@
+import { styled } from '@medly-components/utils';
+
+export const RelativeDiv = styled.div`
+    position: relative;
+`;