+++ packages/core/src/components/CenterAlignedLoader/index.ts
@@ -1,0 +1,1 @@
+export { CenterAlignedLoader as default } from './CenterAlignedLoader';