+++ packages/core/src/components/Table/Foot/index.ts
@@ -1,0 +1,1 @@
+export { Foot as default } from './Foot';