+++ packages/core/src/components/Table/Table.tsx
@@ -4,11 +4,12 @@ import ActionBar from './ActionBar';
 import Body from './Body';
 import ColumnConfiguration from './ColumnConfiguration';
 import { loadingBodyData } from './constants';
+import { TablePropsContext, TableStateContext } from './context';
+import Foot from './Foot';
 import Head from './Head';
 import { getUpdatedColumns } from './helpers';
 import { maxColumnSizeReducer } from './maxColumnSizeReducer';
 import { HiddenDiv, TableStyled } from './Table.styled';
-import { TablePropsContext } from './TableProps.context';
 import { StaticProps, TableProps } from './types';
 import useGroupedRowSelector from './useGroupedRowSelector';
 import useRowSelector from './useRowSelector';
@@ -29,13 +30,19 @@ export const Table: FC<TableProps> & WithStyle & StaticProps = React.memo(
                 isLoading,
                 selectedRowIds,
                 showRowWithCardStyle,
-                enableActionBar,
+                withActionBar,
+                withPagination,
                 ...restProps
             } = props,
             isGroupedTable = !!restProps.groupBy,
             size = showRowWithCardStyle ? 'L' : restProps.size;
 
         const hiddenDivRef = useRef(null),
+            tableState = useState({
+                sortField: props.defaultSortField,
+                sortOrder: props.defaultSortOrder,
+                activePage: props.defaultActivePage
+            }),
             [scrollState, handleScroll] = useScrollState(),
             [selectedGroupIds, setSelectedGroupIds] = useState([]),
             [isSelectAllDisable, setSelectAllDisableState] = useState(true),
@@ -72,44 +79,47 @@ export const Table: FC<TableProps> & WithStyle & StaticProps = React.memo(
         }, [groupedRowSelector.selectedIds]);
 
         return (
-            <TablePropsContext.Provider
-                value={{ ...props, columns, size, data: isLoading ? loadingBodyData : data, isGroupedTable, tableRef, hiddenDivRef }}
-            >
-                <HiddenDiv ref={hiddenDivRef} />
-                {enableActionBar && selectedRowIds.length > 0 && <ActionBar />}
-                <TableStyled
-                    ref={tableRef}
-                    {...restProps}
-                    onScroll={handleScroll}
-                    isRowClickable={isRowClickable}
-                    showRowWithCardStyle={showRowWithCardStyle}
+            <TableStateContext.Provider value={tableState}>
+                <TablePropsContext.Provider
+                    value={{ ...props, columns, size, data: isLoading ? loadingBodyData : data, isGroupedTable, tableRef, hiddenDivRef }}
                 >
-                    <Head
-                        {...{
-                            setColumns,
-                            maxColumnSizes,
-                            areAllRowsSelected,
-                            onSelectAllClick: toggleId,
-                            isSelectAllDisable: isSelectAllDisable,
-                            showShadowAtBottom: !scrollState.isScrolledToTop,
-                            showShadowAfterFrozenElement: !scrollState.isScrolledToLeft,
-                            isAnyRowSelected:
-                                !areAllRowsSelected && (isGroupedTable ? groupedRowSelector.selectedIds.length > 0 : isAnyRowSelected)
-                        }}
-                    />
-                    <Body
-                        {...{
-                            addColumnMaxSize,
-                            setUniqueIds,
-                            setSelectAllDisableState,
-                            selectedRowIds: isGroupedTable ? selectedGroupIds : selectedRowIds,
-                            onRowSelection: toggleId,
-                            onGroupedRowSelection: groupedRowSelector.toggleIds,
-                            showShadowAfterFrozenElement: !scrollState.isScrolledToLeft
-                        }}
-                    />
-                </TableStyled>
-            </TablePropsContext.Provider>
+                    <HiddenDiv ref={hiddenDivRef} />
+                    {withActionBar && selectedRowIds.length > 0 && <ActionBar />}
+                    <TableStyled
+                        {...restProps}
+                        ref={tableRef}
+                        onScroll={handleScroll}
+                        isRowClickable={isRowClickable}
+                        showRowWithCardStyle={showRowWithCardStyle}
+                    >
+                        <Head
+                            {...{
+                                setColumns,
+                                maxColumnSizes,
+                                areAllRowsSelected,
+                                onSelectAllClick: toggleId,
+                                isSelectAllDisable: isSelectAllDisable,
+                                showShadowAtBottom: !scrollState.isScrolledToTop,
+                                showShadowAfterFrozenElement: !scrollState.isScrolledToLeft,
+                                isAnyRowSelected:
+                                    !areAllRowsSelected && (isGroupedTable ? groupedRowSelector.selectedIds.length > 0 : isAnyRowSelected)
+                            }}
+                        />
+                        <Body
+                            {...{
+                                addColumnMaxSize,
+                                setUniqueIds,
+                                setSelectAllDisableState,
+                                selectedRowIds: isGroupedTable ? selectedGroupIds : selectedRowIds,
+                                onRowSelection: toggleId,
+                                onGroupedRowSelection: groupedRowSelector.toggleIds,
+                                showShadowAfterFrozenElement: !scrollState.isScrolledToLeft
+                            }}
+                        />
+                        {withPagination && <Foot />}
+                    </TableStyled>
+                </TablePropsContext.Provider>
+            </TableStateContext.Provider>
         );
     })
 );
@@ -119,9 +129,13 @@ Table.defaultProps = {
     selectedRowIds: [],
     rowIdentifier: 'id',
     defaultSortOrder: 'desc',
+    defaultActivePage: 1,
     rowClickDisableKey: '',
     rowSelectionDisableKey: '',
-    enableActionBar: false
+    withActionBar: false,
+    withPagination: false,
+    totalItems: 0,
+    itemsPerPage: 20
 };
 
 Table.displayName = 'Table';