+++ packages/core/src/components/Table/context/index.ts
@@ -1,0 +1,2 @@
+export * from './TableProps.context';
+export * from './TableState.context';